--- examples/react/vinxi-basic-ssr-streaming/app/routes/no-title.lazy.tsx
@@ -1,14 +1,0 @@
-import { createLazyFileRoute } from '@tanstack/react-router'
-
-export const Route = createLazyFileRoute('/no-title')({
-  component: NoTitle,
-})
-
-function NoTitle() {
-  return (
-    <div>
-      <h1>Hello!</h1>
-      <p>This page has no title.</p>
-    </div>
-  )
-}
