+++ examples/react/vinxi-basic-ssr-streaming/app/routes/no-title.tsx
@@ -1,31 +1,30 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { useEffect, useState, useTransition } from "react";
+import * as React from 'react'
+import { createFileRoute } from '@tanstack/react-router'
 
 async function action() {
-  "use server";
-  console.log("hola from the server");
+  'use server'
+  console.log('hola from the server')
   return new Promise<string>((r) => {
-    setTimeout(() => r("Server says hello, too!"), 500);
-  });
+    setTimeout(() => r('Server says hello, too!'), 500)
+  })
 }
 
-export const Route = createFileRoute("/no-title")({
+export const Route = createFileRoute('/no-title')({
   component: NoTitle,
-});
+})
 
 function NoTitle() {
-  const [isHelloling, startHellling] = useTransition();
-  const [hola, setHola] = useState("");
-  useEffect(() => {
+  const [hola, setHola] = React.useState('')
+  React.useEffect(() => {
     action().then((h) => {
-      setHola(h);
-    });
-  }, []);
+      setHola(h)
+    })
+  }, [])
   return (
     <div>
       {hola}
       <h1>Hello!</h1>
       <p>This page has no title.</p>
     </div>
-  );
+  )
 }
