+++ examples/react/vinxi-basic-ssr-streaming/app/server.tsx
@@ -1,9 +1,13 @@
 /// <reference types="vinxi/types/server" />
+import * as React from 'react'
 import type { PipeableStream } from 'react-dom/server'
 import { renderToPipeableStream } from 'react-dom/server'
 import { eventHandler, toWebRequest, getResponseHeaders } from 'vinxi/server'
 import { getManifest } from 'vinxi/manifest'
-import { StartServer, transformStreamWithRouter } from '@tanstack/react-router-server/server'
+import {
+  StartServer,
+  transformStreamWithRouter,
+} from '@tanstack/react-router-server/server'
 
 import { createRouter } from './router'
 import { createMemoryHistory } from '@tanstack/react-router'
@@ -11,7 +15,6 @@ import {
   serverFnPayloadTypeHeader,
   serverFnReturnTypeHeader,
 } from '@tanstack/react-router-server'
-import React from 'react'
 import { isbot } from 'isbot'
 
 export default eventHandler(async (event) => {
@@ -36,7 +39,7 @@ export default eventHandler(async (event) => {
       {
         tag: 'script',
         children: `window.__vite_plugin_react_preamble_installed__ = true`,
-      }
+      },
     )
   }
 
@@ -82,47 +85,44 @@ export default eventHandler(async (event) => {
 
   const isRobot = isbot(req.headers.get('User-Agent'))
 
-
   const stream = await new Promise<PipeableStream>(async (resolve) => {
     const stream = renderToPipeableStream(<StartServer router={router} />, {
       ...(isRobot
         ? {
-          onAllReady() {
-            resolve(stream)
-          },
-        }
+            onAllReady() {
+              resolve(stream)
+            },
+          }
         : {
-          onShellReady() {
-            resolve(stream)
-          },
-        }),
+            onShellReady() {
+              resolve(stream)
+            },
+          }),
     })
   })
 
   // Add our Router transform to the stream
-  const transforms = [
-    transformStreamWithRouter(router),
-  ]
+  const transforms = [transformStreamWithRouter(router)]
 
   // Pipe the stream through our transforms
   const transformedStream = transforms.reduce(
     (stream, transform) => stream.pipe(transform as any),
-    stream
+    stream,
   )
   ;(event as any).__tsrHeadersSent = true
 
   let headers = {
-      ...getResponseHeaders(event),
-      'Content-Type': 'text/html',
-      ...router.state.matches.reduce((acc, match) => {
-        if (match.headers) {
-          Object.assign(acc, match.headers)
-        }
-        return acc
-      }, {}),
-    }
+    ...getResponseHeaders(event),
+    'Content-Type': 'text/html',
+    ...router.state.matches.reduce((acc, match) => {
+      if (match.headers) {
+        Object.assign(acc, match.headers)
+      }
+      return acc
+    }, {}),
+  }
 
-    // Remove server function headers
+  // Remove server function headers
   ;[serverFnReturnTypeHeader, serverFnPayloadTypeHeader].forEach((header) => {
     delete headers[header]
   })
@@ -137,7 +137,6 @@ export default eventHandler(async (event) => {
     headers,
   })
 })
-
 
 function dedupeHeaders(headerList: [string, string][]): [string, string][] {
   // Object to store the deduplicated headers
